requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Werkzeug==2.3.7
validators==0.22.0
email-validator==2.0.0.post2
Pillow==10.0.0
python-dateutil==2.8.2
zhipuai>=2.0.0
APScheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
markdown==3.4.4
